Det kan potensielt være mange søknader for en person, og requestens levetid ved å hente alle innsynsfiler er ganske høy til tider. Ved å batche opp dette vil det senke levetiden pr. request og data-.overføring pr. request vil være mindre.
